Fix targets to avoid unstable string comparisons
The jsonpb package provides no guarantees about output stability.
Tests that rely on this property are brittle and are unintentionally
becoming a change detector for the jsonpb package, rather than being
a proper test for the code under test. The v2 implementation of jsonpb
will have deliberately unstable outputs to prevent future tests from
relying on this unspecified property.

Fix the test to avoid relying on the exact output of the jsonpb package.

PiperOrigin-RevId: 272293730
Change-Id: Ib0d8d26e2cf6677f48912cfccadcec71d6a70f2a