Token for dataset should not include "identities" field
The error "identities in wrong type" is because we pass a <nil>
to "identities" at hydra accept login. To fix it, we don't pass
"identities" to hydra accept login if "identities" is empty.

Because we always add subject into "identities", it would not be
empty when request a token for endpoint accessing.

PiperOrigin-RevId: 291407341
Change-Id: I600efac852258a1df0e024beda5a5d3f3bbe3643