SCIM link account support
The hydra-ic-test.html page has also been updated to provide a "Link Account" button to provide auth flows to link 2 or more accounts. Also, a new "Account Info" button on this page shows the full account status from the IC, which can be compared and contrasted with what the /userinfo button returns.

PiperOrigin-RevId: 292351245
Change-Id: I39843e0ec7bece0187711a22076abb9211f9b6e1