Update visa translation for older policy engine and verify visa issuers
Uses the config's TrustedPassportIssuers list for trusted visa issuers as well. Renaming that is beyond the scope for this change. DAM config template updated to add persona broker in order to trust visas that it issues.

No support for visas with conditions except in experimental mode. In experimental mode, no support for OR conditions or non-const strings as that was not allowed in the old format.

PiperOrigin-RevId: 295363511
Change-Id: I16e988d226f55bcda1ec877b2f62a7d3a2e88680