Improve deployment and import of configs from files
Each of DAM and IC retain the ability to know how to import config files. It has the ability to accept search/replace variables, but this is only used by tools so far as it assumes that deployed files have already had variables replaced (this can change later as needed). Removes a wipe feature from the services and moves that out to the helper tools, renamed from dam_reset/ic_reset to dam_import/ic_import.

The reset.bash script is renamed to import.bash calling the new import tools with updated parameters.

PiperOrigin-RevId: 296487830
Change-Id: I3670189260c06a9e20f88e2073504aa42fdf6d67