Use same state obj for login and consent process
The issue we can not redirect to app when user click browser back on information release page is because we delete the login state obj when finish login call.

In this fix, we change login and consent process to use same state obj so that we have a challenge to call hydra API to redirect.

PiperOrigin-RevId: 302465737
Change-Id: Iba63dc74a9431de7984423d51a4f6e82dee097df