Add localeMetadata endpoints to IC and DAM
To aid with UIs wanting to set SCIM locale, preferredLanguage, and timezone fields to know what many of the canonical options are to help the users select a reasonable choice. More choices are available, especially for locales, so UI autocomplete dropdowns that allow other values or a similar scheme is recommended.

PiperOrigin-RevId: 321560355
Change-Id: Ie835327384829253270cca7d52ec103baec815ea