Rough in pages for DAM and IC admin guides
These will be used by the UIs to link to more information or help, especially on admin config forms, to get more conceptual information and context.

These docs will be general enough to be reusable across UIs or platform integrations, but may draw on examples from DDAP Explorer or the existing admin UIs. However it is generally assumed that products building on IC/DAM come with their own documentation to help with those specific user journeys.

PiperOrigin-RevId: 322143470
Change-Id: Ic41b061dbd7e009dda51f3434646ccfb71fafd0e