SCIM return string of bad email members
Although the index of the member is already returned as part of a rich error message, include the bad email address string in the error message for use with clients that do not offer a UI experience to correlate the error message to the item.

Also, the SCIM PUT was returning a different error format than a SCIM PATCH because the error message was being rewritten by the request handler layer to drop the rich error structure when detected by the NormalizeInput() method.

PiperOrigin-RevId: 328537972
Change-Id: I3abf6b2193601f8081eb677e87e2f516fff3230a